WIP: Charts page working
